(#11084) Use Puppet 2.7.6 for bootstrap
The bootstrap script seems to use the latest version of puppet
it can find to run through our boostrap modules. This patch simply
switches to the 2.7.6 tag before running puppet apply so we're
processing our modules on the same version we run PE2 with.

Specifically, this works around the issue of Puppet >2.7.6 not
properly creating new file resources with recurse is set to true
on the /root/.vim file resource. That behavior still needs
investigation.