Don't blow away localhost, even if our hostname is different
There was a sed command in the centos.php kickstart script that was replacing
localhost with the machine's domain name. This was bad. Among other things, it
caused console auth to blow up if the machine's official hostname wasn't
"localhost.localdomain."

This commit rearranges it a bit. Now it appends the official hostname to
localhost's entry instead of replacing it; then, once the bootstrap class is
applied, it will split the two entries, so that there's an entry for our
official name and, if our name isn't just localhost, a pristine "127.0.0.1
localhost.localdomain localhost" entry.