IT WORKS!
The edu_bootstrap module now builds out everything I need for the
beta course delivery. To build the environment, declare the
edu_bootstrap::master class in the enterprise console for the
master node and specify the students parameter with a comma
separated list of users to provision. For the agent, simply declare
the edu_bootstrap::agent class in the default group.

The master class builds out user accounts, home directories, module
environments and console accounts. The agent class simply installs
sshfs and creates a folder in roots home for mounting sshfs.

If the beta course delivery goes well, we'll want to add the ability
to automatically provision the environment via eventbrite (though I
bet the demo will go over well!). I'll also want to clean the module
up, document it and publish it on the Forge to puppetlabs namespace
instead of rcoleman.