Polish up edu_bootstrap module
This commit polishes up the edu_bootstrap module to provide a single
interface, the edu_bootstrap class with the students and classes
parameters and adds the ability to programatically add console
classes so that they're in and available for use at the beginning
of class. The version was incremembed to 0.9.0.

TODO - Ryan's on them this week
* Finish documenting the module
* Automatically produce & serve tarballs of user home dirs
* Secure master (chroot homes, manage root password)
* Provide ssh-fuse mount script to agents