Add fuse-sshfs and fix console task for 2.5.2
This commit adds a package resource for fuse-sshfs to the agents
configuration.

It also tweaks the console rake task to be compatible for PE 2.5.2
and sets the role to READ_WRITE (new functionality) instead of the
admin role they previously had. Also, the password is set to the
hash declared at the top of the user manifest (currently puppetlabs).