Fix console password bug
Prior to this commit, I passed the $password hash meant for the unix
user account to the console user creation rake task. This resulted
in a password that is literally the hash.

This commit modifies user.pp to specifiy a separate $console_password
parameter that is used to set it. It defaults to puppetlabs.