Added the modules for the extending capstone
Yes, I committed a binary file. It's an upstream very stable project and
I don't imagine that we'll ever make local changes. Also, it's something
like 2.5 mb uncompressed. Boo.

Anyways.

Classify a node with 'extending::capstone' and you will get:

  * Apache
  * Mysql
  * PHP w/ bindings
  * Drupal & modules I'm using
  * Drupal command line management interface
  * Custom modules to provide:
    * A nifty hiera graphical source:key:value manager
    * A REST API to query values
    * A REST API to post simplified reports viewable from the standard
      Drupal log interface
  * Drupal configured with an admin user. (admin/puppet)

I still have a bit of documentation to work on for the lab. As in a lot.

This requires the packages mentioned in https://projects.puppetlabs.com/issues/18374