Extended to support the Hiera demo
Use Hiera to define the team membership and the team repositories will
all be created and each student will get a working directory for each
team that he/she is a member of.

The Hiera data file should look something like:

---
teams:
  teamone:
    - studentone
    - studenttwo
    - studentthree
  teamtwo:
    - studentfour
    - studentfive

Since the Debian machine should have been spun up by now, it might be
useful for the instructor to add that machine to each team. That will
make it easy to check and/or help out.