Can programatically configure Hiera for Capstone
If you want, you can include the class 'fundamentals::master::hiera' on
the master to automatically generate a `hiera.yaml` and a `common.yaml`
for the Capstone lab. You will have to gauge whether it would be more
helpful to your students to type it out and show how easy it is to
manually edit data files, or to run the code and show them the template
to demonstrate how easy it is to generate programatically.