Add vagrant module
Prior to this commit , we did not add a vagrant user or ssh authorized
key This commit adds a module which adds the user and the key. They key
is the generic insecure public key from vagrant so it works out of the
box. This should only be an issue if people use our VMs in production
which should never happen. Also I suppose is students want to ssh into
each other's vms but that can be done with the same password anyway out
of the box