Significant refactor of the advanced module
This is the WIP refactor. The old module was super hairy to follow and
included a lot of redundant or out of date configuration.

This removes the scripts and whatnot to do the initial setup because
they'll never get used. It also eliminates most of the PuppetDB
configuration, since that's now a first class citizen in PE. It
drastically simplifies how the certificates are managed, so it's finally
easy to see which files go where.

Next up is to clean up the dependencies on external use specific modules
and use the standard issue irc & haproxy modules instead. At that point,
I'll update the version info so we can release a new module.

This is fairly well tested, I banged on it all day. It's still not
ideal, due to the wacky stepped architecture that we're using. I didn't
spend too much time trying to solve that problem though, since I may
just rearchitect the class to a more usual layout.

More eyeballs are appreciated. Please, take a look. Let me know if I
missed anything.