Sync all the certs so agents work again
This syncs all the certs, including private, so that the agents work
again. This is required because when the agents run against themselves,
their master will push out copies of its own certificates... which will
disconnect them from the classroom mesh. This replaces them with the
master certs so that the same certs are synced out.

For the record, I hate this plan. I do not like implicitly promoting it
as a *reasonable* way to do this.