Manages hiera configs for master & agents
This will ensure that hiera is configured on the master so that it
respects the user environments for each student. More details will go
into the instructor guide.

This also configures each student, so they can run puppet apply against
the same hierarchy (wrt their environment, at least).

Finally, it changes how the autoteam bit is calculated, so it can be a
nifty auto data bindings demo, if wanted.