Added simple versioning
This uses a `version.yaml` file to keep track of major/minor versions. I
started this wilth major of 2, because this is the second major iteration
of the build process. I expect that we'll bump that to 3 when we switch
to Packer. The minor version of 1 was arbitrary.

The `/etc/puppetlabs-version` file will contain a version of _[Testing
Build]_ unless the VM is built with the `release` task, which will
automatically increment the minor version and commit it.

Each classroom verification script will now output a version string. My
intent is to add the expected version number to the courseware exercise
guides to keep us honest with our VM release schedule.

This PR is not completely tested due to the difficulty in unit testing
the build process. Please test and tweak as appropriate before merging.