Enables simplified classroom install
This will enable more classroom automation. The only thing it does by
default now is enable time management.

Functionality it provides:
  * students can run a script to automatically configure the classroom.
    It will ask their name and the master's IP and then make magic
    happen. This takes about 45 seconds to complete.
  * instructors can enable $autosetup on the fundamentals class. This
    will manage student's environments and modulepath setting
    automatically. (With the exception of the almost nonexistant case
    where a student is added to multiple teams.)
  * ntp management is enabled by default. The master runs an ntp server
    and all student machines sync from it. The classroom setup script
    automatically syncs, so time problems should never happen again.
  * Pass $offline => true to the fundamentals class and ntp will be
    configured to run in standalone mode.

This is *not* enabled by default, but will be available for
experimentation by instructors willing to troubleshoot and fix breakage.

Based on feedback, this will possibly eliminate most of the classroom
setup labs and somewhat simplify the git workflow lesson.