(COURSES-317) Cached puppetlabs/mysql modules old
The course materials in ciruclation today refer to the new mysql_* types
in the more recent version(s) of the puppetlabs/mysql module. When
performing a class with internet access, there is no issue, as the
Puppet Module Tool can reach the internet. However, when teaching a
class with limited Internet connectivity, the cached version of the
module is too old to be effective.

This commit will set the cached version of the puppetlabs/mysql module
to the 2.3.1, which is the most recent release at this time.

This version has already been tested in the field since this is the
version that would be downloaded in classes with Internet access today.