Disable environment caching for fundamentals
Use the 'environment_timeout' setting to disable environment caching by
setting it to 0.  Since we are rapidly iterating code in the
Fundamentals course, this ensures there's no delay in compiling and
applying the expected catalog.

The default value is "3m", which may cause stale catalogs, delays, and
confusion.

This applies to the 'main' section of puppet.conf.  It should apply to
each student environment as well as the default 'production' environment
for the initial setup of the classroom.