Horrible hack to workaround slow stack start.
The new PE stack takes a very long time to startup, which can cause
disconcerting errors. This simply schedules that to the end of the run
and waits for the service to resume servicing requests before allowing
the run to complete This ensures that if services are restarted, that
will happen *after* any resources that depend on them to be running
are enforced.

I'd appreciate comments from @acidprime @reidmv