Gross hack to let us set user password
This is one of the grossest hacks that we have in the codebase.
It will only exist for a couple weeks until I rewrite all these
defined types as proper API using first class provider citizens.

This allows us to set passwords on new console users so students
can log in.