Fix management of environmentpath
The `puppet_enterprise` module manages only the directory portion of
the `environmentpath`, not any of the contents. We want to put starting
default content in place. Unfortunately, the `environmentpath` is
managed only when classified in the *PE Masters* nodegroup, so we cannot
put contents into it until it's classified.

This introduces a possible parse order problem, but it will fail in the
direction of no action rather than a compile or enforcement failure.

Thanks to @beergeek for identifying the problem.