(re-4124) Package and ship VM artifacts
This adds a task and methods to extract the created VM artifact
from vSphere as an OVA, then ship that OVA to 2 places:

First back into vSphere as a VM suitable for testing.
Second to a directory on the int-resources share.

The shipping tasks need to be run from a machine with
ovftool installed, and if shipping to int-resources, with
the NFS export mounted. Obviously these are intended to be
used on the internal Puppet Labs network.

In addition, there is a minor update required to the Jenkins
build job. The task that exports the VM from vmware needs to know
the name of the pooler VM used. This is accomplished in my test job
by having Beaker write out a properties file containing the name.