Add janky script to make fake reports
Usage: create_nodes.rb [ list, of, nodenames, to, create | remove ]

This script will accept any number of nodenames on the commandline and will
run Puppet in a temporary ssldir for each nodename. This has the effect of
generating an arbitrary number of reports in the Console for screenshotting
purposes. For successful Puppet runs, it will also make a symlink so that
the Puppet Enterprise module can manage mco certificates.

* create_nodes.rb angelina barry craig wilson batman andrew corrine mandy

If you call it with the single argument of `remove` then it will crawl through
the real ssldir and remove all the nasty symlinks we left around.

* create_nodes.rb remove