Add an instructor VM and call it `master`
The `vcmanager` tool expects one machine to be in the `master` role, for indexing purposes. In ID, we don't have that. Each student is standalone. So instead, we'll stand one machine up just like the others that the instructor can use for demos and exercises, but we'll call it the `master`.