Adds a userdata template to go with vcmanager
Goes with https://github.com/puppetlabs/vcmanager/pull/43

If the profile has a `master_ip` set, then this generates a cross platform `#cloud-config` userdata script. The way it works is slightly hacky, but should be solid. The `run-cmd` key applies directly to cloud-init, so it will be executed on the first boot. The `powershell` key is arbitrary, it's just there to make it valid YAML.

EC2Config (the windows service) reads the same file and looks for the `<powershell>` tags. Everything between those tags is evaluated as a PowerShell script on the first boot.