(MAINT) Use legacy branch of pltrainign-bootstrap
Modify the build_files/Puppetfile to use the legacy branch of the
pltraining-bootstrap repository for builds until we can switch entirely
to the new build process implemented in the master branch of that
repository. This will allow us to continue using existing Jenkins build
pipelines unless/until those are adapted and tested against the new
build process.