Update MET AMIs to HVM (#885)
* Update MET AMIs to HVM

Also add a `description` key to the yaml file to help future updates.
This should not affect any tooling, as we don't use that key for
anything.

SDPONBOARD-150 #resolved #time 90m

* Update met.yaml

Convert back to t1.micro on agent instances to get paravirt AMIs working. This allows us to stick with official AMIs for now, even though t1.micro is deprecated in favor of t2.micro. Will have to re-evaluate instance type if t1.micro support is removed in the future.